Move tokenizers into core
