+++ griptape/core/drivers/prompt/cohere_prompt_driver.py
@@ -2,7 +2,7 @@ import cohere
 from attr import define, field, Factory
 from griptape.artifacts import TextOutput
 from griptape.core.drivers import BasePromptDriver
-from griptape.tokenizers import CohereTokenizer
+from griptape.core.tokenizers import CohereTokenizer
 
 
 @define
