+++ tests/unit/tokenizers/test_tiktoken_tokenizer.py
@@ -1,5 +1,5 @@
 import pytest
-from griptape.tokenizers import TiktokenTokenizer
+from griptape.core.tokenizers import TiktokenTokenizer
 
 
 class TestTiktokenTokenizer:
